+++ projects/framework-common-extensions/dialogs/generic_publisher_dialog.py
@@ -156,6 +156,7 @@ class GenericPublisherDialog(BaseContextDialog):
                 "font-style: italic; font-weight: bold;"
             )
             self.tool_widget.layout().addWidget(label_widget)
+            self.run_button.setEnabled(False)
             return
 
         # Create a new tool_config
@@ -388,12 +389,13 @@ class GenericPublisherDialog(BaseContextDialog):
             )
 
     def post_build_ui(self):
-        self._progress_widget.hide_overlay_signal.connect(
-            self.show_main_widget
-        )
-        self._progress_widget.show_overlay_signal.connect(
-            self.show_overlay_widget
-        )
+        if self._progress_widget:
+            self._progress_widget.hide_overlay_signal.connect(
+                self.show_main_widget
+            )
+            self._progress_widget.show_overlay_signal.connect(
+                self.show_overlay_widget
+            )
         self._save_preset_button.clicked.connect(
             self._on_save_preset_button_clicked
         )