+++ projects/framework-houdini/source/ftrack_framework_houdini/__init__.py
@@ -54,7 +54,7 @@ client_instance = None
 
 
 @run_in_main_thread
-def on_run_tool_callback(tool_name, dialog_name=None, options=dict):
+def on_run_tool_callback(tool_name, dialog_name=None, options=None):
     client_instance.run_tool(
         tool_name,
         dialog_name,