+++ projects/framework-photoshop/source/ftrack_framework_photoshop/__init__.py
@@ -166,17 +166,9 @@ def bootstrap_integration(framework_extensions_path):
     registry_instance = registry.Registry()
     registry_instance.scan_extensions(paths=framework_extensions_path)
 
-    Host(
-        event_manager,
-        registry=registry_instance,
-        run_in_main_thread_wrapper=invoke_in_qt_main_thread,
-    )
+    Host(event_manager, registry=registry_instance)
 
-    client_instance = Client(
-        event_manager,
-        registry=registry_instance,
-        run_in_main_thread_wrapper=invoke_in_qt_main_thread,
-    )
+    client_instance = Client(event_manager, registry=registry_instance)
 
     # Init tools
     dcc_config = registry_instance.get_one(