+++ libs/framework-qt/release_notes.md
@@ -3,6 +3,7 @@
 
 ## upcoming
 
+* [new] Add event blocker widget on BaseDialog to prevent unexpected crashes.
 * [changed] BaseDialog; update tool_config_names property to get from options tool_configs key.